+++ rollup.config.js
@@ -13,10 +13,12 @@ const pkg = require('./package.json')
 
 const replaceOptions = {
   exclude: 'node_modules/**',
-  'VERSION': `v${pkg.version}`
+  'process.env.TMLABS_VERSION': JSON.stringify(pkg.version)
 }
 
 if (process.env.production || process.env.browser) replaceOptions['process.env.TMLABS_KEY'] = false
+if (process.env.browser) replaceOptions['process.env.browser'] = true
+else replaceOptions['process.env.browser'] = false
 
 export default {
   entry: 'src/index.js',